pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "personal-news"
version = "0.1.0"
description = "AI-powered daily news digest that fetches, summarizes, and emails personalized news"
authors = [{name = "Personal News", email = "news@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "openai>=1.3.0",
    "anthropic>=0.3.11",
    "beautifulsoup4>=4.12.2",
    "jinja2>=3.1.2",
    "apscheduler>=3.10.4",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "feedparser>=6.0.10",
    "html2text>=2020.1.16",
    "aiohttp>=3.8.5",
    "asyncio-throttle>=1.0.2",
    "google-generativeai>=0.8.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.scripts]
personal-news = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
